updated readme

+++ internal/drones.go
@@ -2,6 +2,6 @@ package internal
 
 //Drone represent a drone
 type Drone struct {
-	ID       string  `json:"id"`
-	SectorID float64 `json:"sectorId"`
+	ID       string `json:"id"`
+	SectorID int    `json:"sectorId"`
 }
