+++ internal/workflow/workflow.go
@@ -17,7 +17,7 @@ func RetrieveLocation(retrieveDrone db.RetrieveDroneByIDFunc) RetrieveLocationFu
 		if err != nil {
 			return pkg.DNSResponse{}, errors.Wrapf(err, "workflow - error retrieving drone details for id=%s", droneID)
 		}
-		sectorID := drone.SectorID
+		sectorID := float64(drone.SectorID)
 		loc := params.XCoord*sectorID + params.YCoord*sectorID + params.ZCoord*sectorID + params.Velocity
 		return pkg.DNSResponse{
 			Location: loc,
